Add comprehensive SEO implementation and client guide
Implement full on-page SEO across all pages: Open Graph tags, Twitter Cards,
JSON-LD structured data (Organization, LocalBusiness, Service, BlogPosting,
Breadcrumb), canonical URLs, and per-page metadata. Split client components
from page files to enable Next.js metadata exports on all routes. Add SEO
guide for client with tool recommendations and monthly checklist.
